+++ src/OpenLoco/Drawing/FPSCounter.h
@@ -1,0 +1,4 @@
+namespace OpenLoco::Drawing
+{
+    void drawFPS();
+}